+++ otter/test/TestOtterIntegration/OtterCore/Fast/Casts/long_long__plus__short_or_char.c
@@ -4,7 +4,7 @@
 #pragma no_other_abandoned
 
 int main(void) {
-    long x;
+    long long x;
     short y;
     char z;
     __SYMBOLIC(&x);
